+++ internal/utilities.go
@@ -1,0 +1,43 @@
+package internal
+
+import (
+	"errors"
+	"reflect"
+	"runtime"
+	"strings"
+)
+
+// Send a function as input param to this function and
+// get the package name of that function as string
+func getPackageName(temp interface{}) string {
+	strs := strings.Split((runtime.FuncForPC(reflect.ValueOf(temp).Pointer()).Name()), ".")
+	strs = strings.Split(strs[len(strs)-2], "/")
+	return strs[len(strs)-1]
+}
+
+// Send a function as input param to this function and
+// get the function name as string
+func getFunctionName(temp interface{}) string {
+	strs := strings.Split((runtime.FuncForPC(reflect.ValueOf(temp).Pointer()).Name()), ".")
+	return strs[len(strs)-1]
+}
+
+// Converts interface to slice
+func InterfaceToSlice(slice interface{}) ([]interface{}, error) {
+	s := reflect.ValueOf(slice)
+	if s.Kind() != reflect.Slice {
+		return nil, errors.New("not a slice")
+	}
+
+	if s.IsNil() {
+		return nil, errors.New("slice is nil")
+	}
+
+	result := make([]interface{}, s.Len())
+
+	for i := 0; i < s.Len(); i++ {
+		result[i] = s.Index(i).Interface()
+	}
+
+	return result, nil
+}