+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
-* ``Chunk`` function added.
+* 🎉 feat(slice-chunk): chunk function added, updated, and optimized
 
 .. 1.0.0 (yyyy-mm-dd)
 .. ------------------